Better image proc, proper qaul1

+++ src/Missions.h
@@ -39,7 +39,7 @@ void missionOverwatchCharlie(int droneId);
 void alphaCorrection(int droneId, int x, int y);
 
 void missionQual1_1(int droneId);
-void missionQual1_2(int droneId, int x, int y);
+void missionQual1_2(int droneId, double x, double y);
 void missionQual1_3(int droneId);
 
 void wait(int droneId);
